requirements.txt
click==8.1.7
Pillow==10.2.0
python-dotenv==1.0.1
torch==2.0.1
transformers==4.31.0
accelerate==0.21.0
coremltools==7.1
sentencepiece==0.1.99 